refactor: restructure Recents module to views-based architecture
Move all Recents-related code to src/views/Recents following the new project structure:

- Move services to src/views/Recents/services/
- fetchRecents, fetchRecentsThunk and all tests
- Create index.ts for proper exports
- Update import paths in storage thunks
